Use -r to require files from bin/disc
Follow the same convention as used in all ruby executables, where you
pass a `-r file` to the executable to require something, instead of
having an environment variable for that.

Particularly, the init file wouldn't change by environment—your workers
will most likely be the same in your local computer, CI, production
server, etc, so it's probably more than reasonable to keep it versioned
in your Procfile or init mechanism, as opposed to an ENV variable.

As a side effect, this allows requiring multiple init files, which I
don't know if it has any real world application, but hey, free feature!

Actually, one real world application is that you can do:

    disc -r celluloid/current -r ./disc_init.rb

And now your code is fully agnostic to whether you're using Celluloid
for parallelizing your workers or not.