Make test-suite pty reads non-blocking
If something goes wrong, the calls to gets can cause the test suite to
hang indefinitely. This changes the code to use non-blocking reads which
will fail immediately if no data is available.

Also bumped the sleep delays, because the original values were to small,
causing above mentioned hangs on slower systems.